+++ UserIgor/XOP_Axon/FP_Mc700Tg/XopUtils.h
@@ -2,7 +2,7 @@
 // XopUtils.h
 
 
-int		 ItemsInList(  char *sString, char *sSep );
+size_t		 ItemsInList(  char *sString, char *sSep );
 void		 StringFromList(  int index, char *sString, char *sSep, char* sItem );
 //char     *StringFromList(  int index, char *sString, char *sSep );
 int       CountSepsInList( char *sString, char *sSep );
